requirements.txt
django
djangorestframework
dj-rest-auth
django-cors-headers
channels
mysqlclient
boto3
requests
Pillow
gunicorn